# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Business Platform Public API",
    "version": "v1",
    "description": "Stable tenant-scoped REST surface for documents, durable processing, governance, and audit. Internal storage keys, raw text, credentials, and provider prompts are never public fields."
  },
  "servers": [{ "url": "http://localhost:3000" }],
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/api/v1/documents": {
      "get": {
        "summary": "List documents",
        "parameters": [{ "$ref": "#/components/parameters/Limit" }, { "$ref": "#/components/parameters/Cursor" }],
        "responses": { "200": { "description": "Document page", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DocumentPageResponse" } } } }, "401": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/api/v1/documents/upload": {
      "post": {
        "summary": "Upload a document and create its metadata",
        "parameters": [{ "$ref": "#/components/parameters/IdempotencyKey" }],
        "requestBody": { "required": true, "content": { "multipart/form-data": { "schema": { "type": "object", "required": ["file"], "properties": { "file": { "type": "string", "format": "binary" } } } } } },
        "responses": { "201": { "description": "Created document", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DocumentResponse" } } } }, "200": { "description": "Idempotent replay", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DocumentResponse" } } } }, "400": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/api/v1/documents/{documentId}": {
      "get": {
        "summary": "Get a document",
        "parameters": [{ "$ref": "#/components/parameters/DocumentId" }],
        "responses": { "200": { "description": "Document", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DocumentResponse" } } } }, "404": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/api/v1/documents/{documentId}/processing-jobs": {
      "get": {
        "summary": "List processing jobs for a document",
        "parameters": [{ "$ref": "#/components/parameters/DocumentId" }],
        "responses": { "200": { "description": "Processing jobs", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProcessingJobArrayResponse" } } } } }
      },
      "post": {
        "summary": "Start durable processing for a content revision",
        "parameters": [{ "$ref": "#/components/parameters/DocumentId" }, { "$ref": "#/components/parameters/IdempotencyKey" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/StartProcessingRequest" } } } },
        "responses": { "201": { "$ref": "#/components/responses/ProcessingJob" }, "200": { "$ref": "#/components/responses/ProcessingJob" }, "409": { "$ref": "#/components/responses/Error" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": { "bearerAuth": { "type": "http", "scheme": "bearer" } },
    "parameters": {
      "Limit": { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 } },
      "Cursor": { "name": "cursor", "in": "query", "schema": { "type": "string", "description": "Versioned opaque cursor" } },
      "IdempotencyKey": { "name": "Idempotency-Key", "in": "header", "required": true, "schema": { "type": "string", "maxLength": 255 } },
      "DocumentId": { "name": "documentId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
...
    },
    "responses": {
      "Error": { "description": "Stable error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
      "ProcessingJob": { "description": "Processing job", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProcessingJobResponse" } } } },
...
    },
    "schemas": {
      "ErrorResponse": { "type": "object", "required": ["code", "message", "request_id"], "properties": { "code": { "type": "string" }, "message": { "type": "string" }, "request_id": { "type": "string" }, "trace_id": { "type": ["string", "null"] }, "details": { "type": ["object", "null"] } } },
      "Document": { "type": "object", "required": ["id", "original_filename", "content_type", "status", "version", "content_revision", "created_at", "updated_at"], "properties": { "id": { "type": "string", "format": "uuid" }, "original_filename": { "type": "string" }, "content_type": { "type": "string" }, "status": { "type": "string" }, "version": { "type": "integer" }, "content_revision": { "type": "integer" }, "size_bytes": { "type": ["integer", "null"] }, "created_at": { "type": "string", "format": "date-time" }, "updated_at": { "type": "string", "format": "date-time" } } },
      "ProcessingJob": { "type": "object", "required": ["job_id", "document_id", "content_revision", "status", "current_step", "attempt_count", "cancel_requested", "candidate_available", "review_available", "created_at", "updated_at"], "properties": { "job_id": { "type": "string", "format": "uuid" }, "document_id": { "type": "string", "format": "uuid" }, "content_revision": { "type": "integer" }, "status": { "type": "string" }, "current_step": { "type": "string" }, "attempt_count": { "type": "integer" }, "failure_code": { "type": ["string", "null"] }, "cancel_requested": { "type": "boolean" }, "candidate_available": { "type": "boolean" }, "review_available": { "type": "boolean" }, "created_at": { "type": "string", "format": "date-time" }, "updated_at": { "type": "string", "format": "date-time" } } },
...
      "StartProcessingRequest": { "type": "object", "required": ["content_revision"], "properties": { "content_revision": { "type": "integer", "minimum": 1 } }, "additionalProperties": false },
...
      "Page": { "type": "object", "required": ["items"], "properties": { "items": { "type": "array" }, "next_cursor": { "type": ["string", "null"] } } },
      "DocumentPage": { "allOf": [{ "$ref": "#/components/schemas/Page" }, { "type": "object", "properties": { "items": { "type": "array", "items": { "$ref": "#/components/schemas/Document" } } } }] },
...
      "ApiEnvelope": { "type": "object", "required": ["success"], "properties": { "success": { "type": "boolean" }, "data": {}, "message": { "type": ["string", "null"] } } },
      "DocumentResponse": { "allOf": [{ "$ref": "#/components/schemas/ApiEnvelope" }, { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/Document" } } }] },
      "ProcessingJobResponse": { "allOf": [{ "$ref": "#/components/schemas/ApiEnvelope" }, { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/ProcessingJob" } } }] },
...
      "DocumentPageResponse": { "allOf": [{ "$ref": "#/components/schemas/ApiEnvelope" }, { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/DocumentPage" } } }] },
...
      "ProcessingJobArrayResponse": { "allOf": [{ "$ref": "#/components/schemas/ApiEnvelope" }, { "type": "object", "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/ProcessingJob" } } } }] },
...
    }
  }
}